refactor signup, register, and forgot password endpoints under /api/v1
