+++ grails-app/views/emails/_email_share_invite.gsp
